requirements.txt
Flask==3.0.3
flask-jwt-extended==4.7.1
python-dotenv==1.0.1
bcrypt==4.2.0
mysql-connector-python==9.0.0
gunicorn==23.0.0
